--- src-ui/sample.ts
@@ -1,3 +1,0 @@
-export const hello: (x: string) => string = x => {
-  return `Hello, ${x}.`;
-};
